requirements.txt
python-telegram-bot == 20.3
fastapi == 0.99.1
python-dotenv == 0.21.1
PyYAML == 6.0
openai == 0.27.8
langchain == 0.0.222